+++ src/components/optimizerTab/OptimizerForm.jsx
@@ -4,22 +4,22 @@ import { Optimizer } from 'lib/optimizer/optimizer'
 import { Constants } from 'lib/constants.ts'
 import { FormRow, OptimizerMenuIds, TeammateFormRow } from 'components/optimizerTab/FormRow.tsx'
 import FormCard from 'components/optimizerTab/FormCard'
-import OptimizerOptionsDisplay from 'components/optimizerTab/OptimizerOptionsDisplay.tsx'
+import OptimizerOptionsDisplay from 'components/optimizerTab/optimizerForm/OptimizerOptionsDisplay.tsx'
 import { OptimizerTabController } from 'lib/optimizerTabController'
 import { SaveState } from 'lib/saveState'
-import { FormSetConditionals } from 'components/optimizerTab/FormSetConditionals'
+import { FormSetConditionals } from 'components/optimizerTab/optimizerForm/FormSetConditionals'
 import DB from 'lib/db'
 import { Utils } from 'lib/utils.js'
-import { CharacterConditionalDisplay } from 'components/optimizerForm/conditionals/CharacterConditionalDisplay'
-import { LightConeConditionalDisplay } from 'components/optimizerForm/conditionals/LightConeConditionalDisplay'
-import TeammateCard from 'components/optimizerTab/TeammateCard'
-import CharacterSelectorDisplay from 'components/optimizerTab/CharacterSelectorDisplay.tsx'
-import EnemyOptionsDisplay from 'components/optimizerTab/EnemyOptionsDisplay'
-import RelicMainSetFilters from 'components/optimizerTab/RelicMainSetFilters'
-import { SubstatWeightFilters } from 'components/optimizerTab/SubstatWeightFilters'
-import { MinMaxRatingFilters, MinMaxStatFilters } from 'components/optimizerTab/ResultFilters'
-import { CombatBuffsFilters } from 'components/optimizerTab/CombatBuffsFilters'
-import { OptimizerTabCharacterPanel } from 'components/optimizerTab/OptimizerTabCharacterPanel'
+import { CharacterConditionalDisplay } from 'components/optimizerTab/conditionals/CharacterConditionalDisplay'
+import { LightConeConditionalDisplay } from 'components/optimizerTab/conditionals/LightConeConditionalDisplay'
+import TeammateCard from 'components/optimizerTab/optimizerForm/TeammateCard'
+import CharacterSelectorDisplay from 'components/optimizerTab/optimizerForm/CharacterSelectorDisplay.tsx'
+import EnemyOptionsDisplay from 'components/optimizerTab/optimizerForm/EnemyOptionsDisplay'
+import RelicMainSetFilters from 'components/optimizerTab/optimizerForm/RelicMainSetFilters'
+import { SubstatWeightFilters } from 'components/optimizerTab/optimizerForm/SubstatWeightFilters'
+import { MinMaxRatingFilters, MinMaxStatFilters } from 'components/optimizerTab/optimizerForm/ResultFilters'
+import { CombatBuffsFilters } from 'components/optimizerTab/optimizerForm/CombatBuffsFilters'
+import { OptimizerTabCharacterPanel } from 'components/optimizerTab/optimizerForm/OptimizerTabCharacterPanel'
 import { LightConeConditionals } from 'lib/lightConeConditionals'
 import FilterContainer from 'components/optimizerTab/FilterContainer.tsx'
