+++ src/components/optimizerTab/conditionals/CharacterConditionalDisplay.tsx
@@ -1,10 +1,10 @@
 import { Flex } from 'antd'
 import { memo } from 'react'
-import { HeaderText } from 'components/HeaderText'
-import { TooltipImage } from 'components/TooltipImage'
-import { Hint } from 'lib/hint'
-import DisplayFormControl from './DisplayFormControl'
-import { characterOptionMapping } from 'lib/characterConditionals'
+import { HeaderText } from 'components/HeaderText.jsx'
+import { TooltipImage } from 'components/TooltipImage.jsx'
+import { Hint } from 'lib/hint.jsx'
+import DisplayFormControl from 'components/optimizerTab/conditionals/DisplayFormControl.tsx'
+import { characterOptionMapping } from 'lib/characterConditionals.js'
 import { Eidolon } from 'types/Character'
 import { DataMineId } from 'types/Common'
