+++ src/components/optimizerTab/conditionals/LightConeConditionalDisplay.tsx
@@ -1,12 +1,12 @@
 import { Flex, Typography } from 'antd'
-import { HeaderText } from 'components/HeaderText'
-import { TooltipImage } from 'components/TooltipImage'
-import { Hint } from 'lib/hint'
+import { HeaderText } from 'components/HeaderText.jsx'
+import { TooltipImage } from 'components/TooltipImage.jsx'
+import { Hint } from 'lib/hint.jsx'
 import { DataMineId } from 'types/Common'
 import { SuperImpositionLevel } from 'types/LightCone'
 import { memo } from 'react'
-import { lightConeOptionMapping } from 'lib/lightConeConditionals'
-import DisplayFormControl from './DisplayFormControl.tsx'
+import { lightConeOptionMapping } from 'lib/lightConeConditionals.js'
+import DisplayFormControl from 'components/optimizerTab/conditionals/DisplayFormControl.tsx'
 
 export interface LightConeConditionalDisplayProps {
   id?: DataMineId
