+++ src/components/optimizerTab/optimizerForm/CharacterSelectorDisplay.tsx
@@ -1,11 +1,11 @@
 import { Flex, Form, Select } from 'antd'
-import { HeaderText } from 'components/HeaderText'
-import { TooltipImage } from 'components/TooltipImage'
-import { Hint } from 'lib/hint'
-import { Utils } from 'lib/utils'
-import { eidolonOptions, levelOptions, superimpositionOptions } from 'lib/constants'
-import RecommendedPresetsButton from 'components/optimizerForm/RecommendedPresetsButton'
-import { optimizerTabDefaultGap, panelWidth } from 'components/optimizerTab/optimizerTabConstants'
+import { HeaderText } from 'components/HeaderText.jsx'
+import { TooltipImage } from 'components/TooltipImage.jsx'
+import { Hint } from 'lib/hint.jsx'
+import { Utils } from 'lib/utils.js'
+import { eidolonOptions, levelOptions, superimpositionOptions } from 'lib/constants.ts'
+import RecommendedPresetsButton from 'components/optimizerTab/optimizerForm/RecommendedPresetsButton.tsx'
+import { optimizerTabDefaultGap, panelWidth } from 'components/optimizerTab/optimizerTabConstants.ts'
 import { useEffect, useMemo } from 'react'
 import { OptimizerTabController } from 'lib/optimizerTabController.js'
