+++ src/components/optimizerTab/optimizerForm/OptimizerOptionsDisplay.tsx
@@ -3,8 +3,8 @@ import { Flex, Form, Select, Switch, Typography } from 'antd'
 import { CheckOutlined, CloseOutlined } from '@ant-design/icons'
 
 import { Hint } from 'lib/hint.jsx'
-import { HeaderText } from '../HeaderText'
-import { TooltipImage } from '../TooltipImage'
+import { HeaderText } from 'components/HeaderText.jsx'
+import { TooltipImage } from 'components/TooltipImage.jsx'
 import { useMemo } from 'react'
 import DB from 'lib/db.js'
 import { optimizerTabDefaultGap, panelWidth } from 'components/optimizerTab/optimizerTabConstants.ts'
