+++ src/components/optimizerTab/optimizerForm/RelicMainSetFilters.tsx
@@ -4,8 +4,8 @@ import { TooltipImage } from 'components/TooltipImage.jsx'
 import { Hint } from 'lib/hint.jsx'
 import { optimizerTabDefaultGap, panelWidth } from 'components/optimizerTab/optimizerTabConstants.ts'
 import { Constants, RelicSetFilterOptions } from 'lib/constants.ts'
-import GenerateSetsOptions from 'components/optimizerForm/SetsOptions.tsx'
-import GenerateOrnamentsOptions from 'components/optimizerForm/OrnamentsOptions.tsx'
+import GenerateSetsOptions from 'components/optimizerTab/optimizerForm/SetsOptions.tsx'
+import GenerateOrnamentsOptions from 'components/optimizerTab/optimizerForm/OrnamentsOptions.tsx'
 import { SettingOutlined } from '@ant-design/icons'
 import { Assets } from 'lib/assets.js'
 import PropTypes from 'prop-types'
