+++ src/components/optimizerTab/optimizerForm/SubstatWeightFilters.tsx
@@ -2,7 +2,10 @@ import { Flex, Typography } from 'antd'
 import { HeaderText } from 'components/HeaderText.jsx'
 import { TooltipImage } from 'components/TooltipImage.jsx'
 import { Hint } from 'lib/hint.jsx'
-import { FormStatRollSlider, FormStatRollSliderTopPercent } from 'components/optimizerTab/FormStatRollSlider.jsx'
+import {
+  FormStatRollSlider,
+  FormStatRollSliderTopPercent,
+} from 'components/optimizerTab/optimizerForm/FormStatRollSlider.jsx'
 import { Constants } from 'lib/constants.ts'
 import { HorizontalDivider } from 'components/Dividers.tsx'
 import { optimizerTabDefaultGap } from 'components/optimizerTab/optimizerTabConstants.ts'
