+++ src/components/optimizerTab/optimizerForm/TeammateCard.tsx
@@ -8,8 +8,8 @@ import { Assets } from 'lib/assets.js'
 import { CharacterConditionals } from 'lib/characterConditionals.js'
 import { LightConeConditionals } from 'lib/lightConeConditionals.js'
 import { OptimizerTabController } from 'lib/optimizerTabController.js'
-import { CharacterConditionalDisplay } from 'components/optimizerForm/conditionals/CharacterConditionalDisplay.tsx'
-import { LightConeConditionalDisplay } from 'components/optimizerForm/conditionals/LightConeConditionalDisplay.tsx'
+import { CharacterConditionalDisplay } from 'components/optimizerTab/conditionals/CharacterConditionalDisplay.tsx'
+import { LightConeConditionalDisplay } from 'components/optimizerTab/conditionals/LightConeConditionalDisplay.tsx'
 import DB from 'lib/db.js'
 import { Character } from 'types/Character'
 import { Message } from 'lib/message.js'
